fix grunt server not running compile and watch
using async was a bad idea here. this way, the other tasks will now run after
the process is spawned instead of waiting for it to fail. by using
grunt.fail.fatal in the callback we ensure a proper message is displayed to the
user in case nodemon itself crashes